+++ src/core/object/proxy-clone/const.ts
@@ -6,6 +6,11 @@
  * https://github.com/V4Fire/Core/blob/master/LICENSE
  */
 
+import { PROXY } from 'core/prelude/types/const';
+
 export const
 	SELF = {},
 	NULL = {};
+
+export const
+	toOriginalObject = PROXY;