+++ src/core/object/proxy-clone/index.ts
@@ -12,14 +12,13 @@
  */
 
 import { unimplement } from 'core/functools/implementation';
-import { NULL } from 'core/object/proxy-clone/const';
-import { resolveTarget } from 'core/object/proxy-clone/helpers';
 
-export * from 'core/object/proxy-clone/const';
+import { toOriginalObject, NULL } from 'core/object/proxy-clone/const';
+import { resolveTarget, getRawValueFromStore, Descriptor } from 'core/object/proxy-clone/helpers';
 
 /**
  * Returns a clone of the specified object.
- * To clone an object, the function creates a Proxy object, i.e., the process of cloning is a lazy operation.
+ * The function uses a Proxy object to create a clone. The process of cloning is a lazy operation.
  *
  * @param obj
  */
@@ -34,45 +33,50 @@ export default function proxyClone<T>(obj: T): T {
 
 		if (typeof Proxy !== 'function') {
 			unimplement({
-				name: 'clone',
+				name: 'proxyClone',
 				type: 'function',
 				notice: 'An operation of proxy object cloning depends on the support of native Proxy API'
 			});
 		}
 
+		let
+			lastSetKey;
+
 		const proxy = new Proxy(Object.cast<object>(obj), {
 			get: (target, key, receiver) => {
-				const originalTarget = target;
-				target = resolveTarget(target, store).value;
-
-				let
-					valStore = store.get(target);
-
-				let
-					val;
-
-				if (valStore?.has(key)) {
-					val = valStore.get(key);
-					val = val === NULL ? undefined : val;
-
-				} else {
-					val = Reflect.get(target, key, receiver);
+				if (key === toOriginalObject) {
+					return target;
 				}
 
 				const
-					isArray = Object.isArray(target),
-					isCustomObject = isArray || Object.isCustomObject(target);
+					resolvedTarget = resolveTarget(target, store).value;
 
-				if (isArray && !Reflect.has(target, Symbol.isConcatSpreadable)) {
-					target[Symbol.isConcatSpreadable] = true;
+				let
+					valStore = store.get(resolvedTarget),
+					val = getRawValueFromStore(key, valStore) ?? Reflect.get(resolvedTarget, key, receiver);
+
+				if (val instanceof Descriptor) {
+					val = val.getValue(receiver);
+				}
+
+				if (val === NULL) {
+					val = undefined;
+				}
+
+				const
+					isArray = Object.isArray(resolvedTarget),
+					isCustomObject = isArray || Object.isCustomObject(resolvedTarget);
+
+				if (isArray && !Reflect.has(resolvedTarget, Symbol.isConcatSpreadable)) {
+					resolvedTarget[Symbol.isConcatSpreadable] = true;
 				}
 
 				if (Object.isFunction(val) && !isCustomObject) {
-					if (Object.isMap(target) || Object.isSet(target)) {
+					if (Object.isMap(resolvedTarget) || Object.isSet(resolvedTarget)) {
 						switch (key) {
 							case 'get':
 								return (...args) => {
-									const val = target[key](...args);
+									const val = resolvedTarget[key](...args);
 									return clone(val);
 								};
 
@@ -81,7 +85,7 @@ export default function proxyClone<T>(obj: T): T {
 							case Symbol.iterator:
 								return (...args) => {
 									const
-										iter = target[key](...args);
+										iter = resolvedTarget[key](...args);
 
 									return {
 										[Symbol.iterator]() {
@@ -101,19 +105,19 @@ export default function proxyClone<T>(obj: T): T {
 								};
 
 							default:
-								return val.bind(target);
+								return val.bind(resolvedTarget);
 						}
 					}
 
-					if (Object.isWeakMap(target) || Object.isWeakSet(target)) {
+					if (Object.isWeakMap(resolvedTarget) || Object.isWeakSet(resolvedTarget)) {
 						switch (key) {
 							case 'get':
 								return (prop) => {
 									if (valStore == null || !valStore.has(prop)) {
-										return clone(Object.cast<Map<unknown, unknown>>(originalTarget).get(prop));
+										return clone(Object.cast<Map<unknown, unknown>>(target).get(prop));
 									}
 
-									const val = target[key](prop);
+									const val = resolvedTarget[key](prop);
 									return clone(val);
 								};
 
@@ -121,44 +125,57 @@ export default function proxyClone<T>(obj: T): T {
 								return (prop, val) => {
 									valStore ??= new Map();
 									valStore.set(prop, val);
-									return target[key](prop, val);
+									return resolvedTarget[key](prop, val);
 								};
 
 							case 'add':
 								return (prop) => {
 									valStore ??= new Map();
 									valStore.set(prop, true);
-									return target[key](prop);
+									return resolvedTarget[key](prop);
 								};
 
 							case 'has':
 								return (prop) => {
 									if (valStore == null || !valStore.has(prop)) {
-										return Object.cast<Set<unknown>>(originalTarget).has(prop);
+										return Object.cast<Set<unknown>>(target).has(prop);
 									}
 
-									return target[key](prop);
+									return resolvedTarget[key](prop);
 								};
 
 							default:
-								return val.bind(target);
+								return val.bind(resolvedTarget);
 						}
 					}
 
-					return val.bind(target);
+					return val.bind(resolvedTarget);
 				}
 
 				return clone(val);
 			},
 
 			set: (target, key, val, receiver) => {
+				lastSetKey = key;
+
 				const {
 					value: resolvedTarget,
 					needWrap
 				} = resolveTarget(target, store);
 
+				if (!needWrap) {
+					return Reflect.set(resolvedTarget, key, val);
+				}
+
 				const
-					desc = Reflect.getOwnPropertyDescriptor(resolvedTarget, key);
+					rawValue = getRawValueFromStore(key, store.get(resolvedTarget));
+
+				if (rawValue instanceof Descriptor) {
+					return rawValue.setValue(val, receiver);
+				}
+
+				const
+					desc = Reflect.getOwnPropertyDescriptor(receiver, key);
 
 				if (desc != null) {
 					if (desc.writable === false || desc.get != null && desc.set == null) {
@@ -171,17 +188,106 @@ export default function proxyClone<T>(obj: T): T {
 					}
 				}
 
-				if (needWrap) {
-					const
-						valStore = store.get(resolvedTarget) ?? new Map();
-
+				if ((key in resolvedTarget)) {
+					const valStore = store.get(resolvedTarget) ?? new Map();
 					store.set(resolvedTarget, valStore);
 					valStore.set(key, val);
 
-					return true;
+				} else {
+					Object.defineProperty(receiver, key, {
+						configurable: true,
+						writable: true,
+						enumerable: true,
+						value: val
+					});
 				}
 
-				return Reflect.set(resolvedTarget, key, val);
+				return true;
+			},
+
+			defineProperty: (target, key, desc) => {
+				const {
+					value: resolvedTarget,
+					needWrap
+				} = resolveTarget(target, store);
+
+				if (!needWrap || lastSetKey === key) {
+					if (lastSetKey === key) {
+						lastSetKey = undefined;
+					}
+
+					return Reflect.defineProperty(resolvedTarget, key, desc);
+				}
+
+				const
+					rawValue = getRawValueFromStore(key, store.get(resolvedTarget)),
+					oldDesc = Reflect.getOwnPropertyDescriptor(resolvedTarget, key);
+
+				if (
+					oldDesc?.configurable === false &&
+					(oldDesc.writable === false || Object.size(desc) > 1 || !('value' in desc))
+				) {
+					return false;
+				}
+
+				const mergedDesc = {
+					configurable: desc.configurable !== false
+				};
+
+				if (oldDesc != null) {
+					const baseDesc: PropertyDescriptor = {
+						configurable: oldDesc.configurable,
+						enumerable: oldDesc.enumerable
+					};
+
+					Object.assign(mergedDesc, baseDesc);
+
+					if (rawValue instanceof Descriptor) {
+						Object.assign(mergedDesc, rawValue.descriptor);
+					}
+
+					Object.assign(mergedDesc, desc);
+
+					if (desc.get != null || desc.set != null) {
+						delete mergedDesc['value'];
+						delete mergedDesc['writable'];
+
+					} else {
+						delete mergedDesc['get'];
+						delete mergedDesc['set'];
+					}
+
+				} else {
+					Object.assign(mergedDesc, desc);
+				}
+
+				const
+					valStore = store.get(resolvedTarget) ?? new Map();
+
+				store.set(resolvedTarget, valStore);
+				valStore.set(key, new Descriptor(mergedDesc));
+
+				if (!(key in resolvedTarget)) {
+					Object.defineProperty(resolvedTarget, key, {
+						configurable: true,
+						enumerable: false,
+
+						set: (value) => {
+							Object.defineProperty(resolvedTarget, key, {
+								enumerable: true,
+								writable: true,
+								configurable: true,
+								value
+							});
+						},
+
+						get() {
+							return undefined;
+						}
+					});
+				}
+
+				return true;
 			},
 
 			deleteProperty: (target, key) => {
@@ -190,31 +296,41 @@ export default function proxyClone<T>(obj: T): T {
 					needWrap
 				} = resolveTarget(target, store);
 
+				if (!needWrap) {
+					return Reflect.deleteProperty(resolvedTarget, key);
+				}
+
 				const
 					desc = Reflect.getOwnPropertyDescriptor(resolvedTarget, key);
 
-				if (desc != null) {
-					if (desc.writable === false || desc.get != null && desc.set == null) {
-						return false;
-					}
+				if (desc?.configurable === false) {
+					return false;
 				}
 
-				if (needWrap) {
-					const
-						valStore = store.get(resolvedTarget) ?? new Map();
+				const
+					valStore = store.get(resolvedTarget) ?? new Map();
 
-					store.set(resolvedTarget, valStore);
-					valStore.set(key, NULL);
+				store.set(resolvedTarget, valStore);
+				valStore.set(key, NULL);
 
-					return true;
-				}
-
-				return Reflect.deleteProperty(target, key);
+				return true;
 			},
 
 			has: (target, key) => {
+				if (key === toOriginalObject) {
+					return true;
+				}
+
+				const {
+					value: resolvedTarget,
+					needWrap
+				} = resolveTarget(target, store);
+
+				if (!needWrap) {
+					return Reflect.has(resolvedTarget, key);
+				}
+
 				const
-					resolvedTarget = resolveTarget(target, store).value,
 					valStore = store.get(resolvedTarget);
 
 				if (valStore?.has(key)) {
@@ -222,7 +338,71 @@ export default function proxyClone<T>(obj: T): T {
 				}
 
 				return Reflect.has(resolvedTarget, key);
-			}
+			},
+
+			getOwnPropertyDescriptor: (target, key) => {
+				const {
+					value: resolvedTarget,
+					needWrap
+				} = resolveTarget(target, store);
+
+				const
+					desc = Reflect.getOwnPropertyDescriptor(resolvedTarget, key);
+
+				if (!needWrap || desc == null) {
+					return desc;
+				}
+
+				const
+					rawVal = getRawValueFromStore(key, store.get(resolvedTarget));
+
+				if (rawVal instanceof Descriptor) {
+					const
+						rawDesc = rawVal.descriptor;
+
+					if (desc.configurable) {
+						return rawVal.descriptor;
+					}
+
+					const
+						mergedDesc = {...desc};
+
+					if (rawDesc.get == null && rawDesc.set == null) {
+						mergedDesc.value = rawVal.getValue(proxy);
+					}
+
+					return mergedDesc;
+				}
+
+				return desc;
+			},
+
+			ownKeys: (target) => {
+				const {
+					value: resolvedTarget,
+					needWrap
+				} = resolveTarget(target, store);
+
+				if (!needWrap) {
+					return Reflect.ownKeys(resolvedTarget);
+				}
+
+				const
+					keys = new Set(Reflect.ownKeys(resolvedTarget));
+
+				Object.forEach(store.get(resolvedTarget)?.entries(), ([key, val]) => {
+					if (val === NULL) {
+						keys.delete(key);
+
+					} else if (key in resolvedTarget) {
+						keys.add(key);
+					}
+				});
+
+				return [...keys];
+			},
+
+			preventExtensions: () => false
 		});
 
 		return Object.cast(proxy);