+++ src/core/request/response/index.ts
@@ -629,20 +629,24 @@ export default class Response<
 			});
 		});
 
-		res = res.then((res) => {
-			if (Object.isFrozen(res)) {
-				return res;
+		res = res.then((data) => {
+			if (Object.isFrozen(data)) {
+				return data;
 			}
 
-			if (Object.isArray(res) || Object.isPlainObject(res)) {
-				Object.defineProperty(res, 'valueOf', {
-					value: () => Object.fastClone(res, {freezable: false})
+			if (Object.isArray(data) || Object.isPlainObject(data)) {
+				const
+					originalData = data;
+
+				Object.defineProperty(data, 'valueOf', {
+					configurable: true,
+					value: () => Object.fastClone(originalData, {freezable: false})
 				});
 
-				Object.freeze(res);
+				Object.freeze(data);
 			}
 
-			return res;
+			return data;
 		});
 
 		return Object.cast(res);