+++ src/core/request/spec.js
@@ -475,7 +475,7 @@ fdescribe('core/request', () => {
 				await retryDelayTest(() => new Promise((res) => setTimeout(res, 200)), 200);
 			});
 
-			it('retrying with the speeduped response', async () => {
+			it('retrying with the speedup response', async () => {
 				const req = await request('http://localhost:3000/retry/speedup', {
 					timeout: 300,
 					retry: 2
@@ -691,8 +691,10 @@ fdescribe('core/request', () => {
 
 				const
 					body = await (await req).response.json(),
-					firstRequest = body.times.shift(),
-					requestDelays = body.times.reduce((acc, time, i) => acc.concat(time - firstRequest - i * delayMS), []);
+					firstRequest = body.times[0];
+
+				const requestDelays = body.times.slice(1)
+					.reduce((acc, time, i) => acc.concat(time - firstRequest - i * delayMS), []);
 
 				expect(firstRequest - startTime).toBeLessThan(delayMS);
 				requestDelays.forEach((time) => expect(time).toBeGreaterThanOrEqual(delayMS));